- Added casts to better match other branches

+++ PNGParser/PNGParser.cpp
@@ -334,7 +334,7 @@ std::vector<Byte> DecompressImage(std::vector<Byte> dataBytes, const ChunkData<"
 {
     z_stream zstream = {};
     zstream.next_in = dataBytes.data();
-    zstream.avail_in = dataBytes.size();
+    zstream.avail_in = static_cast<std::uint32_t>(dataBytes.size());
 
     if(inflateInit(&zstream) != Z_OK)
         throw std::exception("zstream failed to initialize");
@@ -351,7 +351,7 @@ std::vector<Byte> DecompressImage(std::vector<Byte> dataBytes, const ChunkData<"
     decompressedImage.resize(DecompressedImageSize(headerData));
 
     zstream.next_out = &decompressedImage[0];
-    zstream.avail_out = decompressedImage.size();
+    zstream.avail_out = static_cast<std::uint32_t>(decompressedImage.size());
 
     if(auto cont = inflate(&zstream, Z_FULL_FLUSH); !(cont == Z_OK || cont == Z_STREAM_END))
     {
